feat: 이미지 캐러셀 훅
